requirements.txt
absl-py==2.1.0
attrs==25.1.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
faiss-cpu==1.10.0
filelock==3.13.1
fonttools==4.56.0
fsspec==2024.6.1
fvcore==0.1.5.post20221221
grpcio==1.70.0
h11==0.14.0
hdbscan==0.8.40
idna==3.10
iopath==0.1.10
Jinja2==3.1.3
joblib==1.4.2
kiwisolver==1.4.8
Markdown==3.7
MarkupSafe==2.1.5
matplotlib==3.10.0
mpmath==1.3.0
networkx==3.3
numpy==2.2.2
opencv-python==4.11.0.86
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
pillow==11.1.0
portalocker==3.1.1
protobuf==5.29.3
pycparser==2.22
pyparsing==3.2.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2025.1
pywin32==308
PyYAML==6.0.2
requests==2.32.3
scikit-learn==1.6.1
scipy==1.15.1
seaborn==0.13.2
selenium==4.28.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
sympy==1.13.1
tabulate==0.9.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
termcolor==2.5.0
threadpoolctl==3.5.0
torch==2.6.0+cu126
torchaudio==2.6.0+cu126
torchvision==0.21.0+cu126
tqdm==4.67.1
trio==0.28.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
websocket-client==1.8.0
Werkzeug==3.1.3
wsproto==1.2.0
yacs==0.1.8
